requirements.txt
azure-functions>=1.15.0
azure-storage-blob>=12.14.0
azure-mgmt-datafactory>=1.1.0
azure-mgmt-sql>=3.0.0
pyodbc>=4.0.34
databricks-connect>=13.0.0
apache-airflow>=2.7.0
pandas>=1.5.0
numpy>=1.23.0
pytest>=7.4.0
python-dotenv>=1.0.0
great-expectations>=0.15.0
pylint>=2.17.0
black>=23.7.0
pyspark>=3.3.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
networkx>=3.0
pydot>=1.4.2
PyYAML>=6.0
requests>=2.28.0
python-dateutil>=2.8.2
typing-extensions>=4.4.0
dataclasses>=0.6
